requirements.txt
fastapi
langchain-aws
uvicorn
boto3
langchain

# dev dependencies
python-dotenv